Read SysEx events correctly
SysExStart is followed by a length, which should be taken into account.
Ending on SysExEnd is still expected, but the length isn't part of the
message data.

SysExEnd i.e. 0xf7 can also show up as a message in its own right, this
is an "escaped" sysex event or a "continuation" of a longer sysex event.
In that case the 0xf7 is not part of the data.